FIX: Restore ``AFNICommand._get_fname``, required by some interfaces
In #2964 I broke some interfaces (@tsalo identified `Allineate` as one
of those) by removing the ``_get_fname`` function from the base
``AFNICommand``.

The intent was to migrate to a ``name_source``-based management of
automatically generated names, but in the case of ``Allineate``, the use
of `_get_fname` was a bit different (modifying the user-provide input
value under certain conditions to follow ANFI's ``3dAllineate``
behavior).

Also remove a definition of the same function for ``Deconvolve``.

This PR restores the missing function. Closes #3070.
